% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_procedure.R
\name{create_procedure}
\alias{create_procedure}
\title{Create Procedure to accept and store a model}
\usage{
create_procedure(connection, procedure_name, table_name, ..)
}
\arguments{
\item{connection}{a connection to the database from create connection}

\item{procedure_name}{name of the procedure to create}

\item{table_name}{name to the table to store the model}
}
\description{
Model Deployment
}
\examples{
\dontrun{
# don't run this sript
}
con = create_connection(server = "",database = "",username="",password="")
create_procedure(con,"model_procedure","model_table")
}
\seealso{
\code{\link{createConnection}}
}
