% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createConnection.R
\name{createConnection}
\alias{createConnection}
\title{Create a connection to the server}
\usage{
createConnection(server, database, username, password, ...)
}
\arguments{
\item{server}{server name for authentication (charcter string)}

\item{database}{database name}

\item{username}{username for authentication}

\item{password}{for authenitication (if required)}
}
\description{
Create a connection to the server
}
\examples{
\dontrun{
# don't run this sript
}
con = create_connection(server = "",database = "",username="",password="")
}
\seealso{
\code{\link[RODBC]{odbcConnect}}
}
