% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_procedure.R
\name{createProcedure}
\alias{createProcedure}
\title{Create Procedure to accept and store a model}
\usage{
createProcedure(connection, procedure_name, table_name)
}
\arguments{
\item{connection}{a connection to the database from createConnection}

\item{procedure_name}{name of the procedure to create}

\item{table_name}{name of the table to accept/store the model}
}
\description{
Model Deployment
}
\examples{
\dontrun{
# don't run this sript
con = createConnection(server = "",database = "",username="",password="")
createProcedure(con,"model_procedure","model_table")
}
}
\seealso{
\code{\link{createConnection}}
}
