% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployModel.R
\name{deployModel}
\alias{deployModel}
\title{Deploy Model to SQL Server, Spark And Hadoop}
\usage{
deployModel(connection, procedure_name, model, name, type, description,
  version, author)
}
\arguments{
\item{connection}{RODBC connection object returned by createConnection}

\item{procedure_name}{name of the procedure as passed to create_procedure}

\item{model}{model object as returned by serialize_model}

\item{name}{name of the algorithm. example: Logistic Regression model, LightGBM model}

\item{type}{model type (classification or Regression, et.c)}

\item{description}{model description}

\item{version}{model version}

\item{author}{name of the author/creator}
}
\description{
Deploy Model to SQL Server, Spark And Hadoop
}
\examples{
\dontrun{
# don't run this sript
}
con = create_connection(server = "",database = "",username="",password="")
}
\seealso{
\code{\link{createConnection} \link{create_procedure} \link{serializeModel}}
}
