% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializeModel.R
\name{serializeModel}
\alias{serializeModel}
\title{Simple Model Serialization Interface}
\usage{
serializeModel(model)
}
\arguments{
\item{model}{model object}
}
\value{
A binary output
}
\description{
A simple low-level interface for serializing models.
}
\examples{
\dontrun{
# don't run this sript
}
x <- serialize(list(1,2,3), NULL)
unserialize(x)
}
\seealso{
\code{\link{serialize} \link{unserialize}}
}
