% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripModel.R
\name{stripModel}
\alias{stripModel}
\title{Strip and Remove unneccesary large files associated with GLM and LM models}
\usage{
stripModel(model)
}
\arguments{
\item{model}{a GLM or LM model}
}
\description{
Strip and Remove unneccesary large files associated with models. The function helps reduces the size of the model.
}
\examples{
\dontrun{
# don't run this sript
}
con = create_connection(server = "",database = "",username="",password="")
}
